pyproject.toml
[tool.poetry]
name = "pytest-fixture-maker"
version = "0.1.0"
description = "Pytest plugin to load fixtures from YAML files"
authors = ["Kirill Sumorokov <sumorokov.k@gmail.com>"]
packages = [
    { include = "pytest_fixture_maker", from = "src" }
]

[tool.poetry.plugins.pytest11]
fixture_maker = "pytest_fixture_maker.plugin"

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
black = "19.10b0"
coverage = "5.5"
faker = "8.8.2"
flake8 = "3.9.2"
flake8-bugbear = "21.4.3"
flake8-builtins-unleashed = "1.3.1"
flake8-comprehensions = "3.5.0"
flake8-import-order = "0.18.1"
flake8-print = "4.0.0"
ipython = "7.25.0"
mypy = "0.910"
pytest = "^6.2.5"
pytest-cov = "2.12.1"
pytest-flake8 = "1.0.7"
pytest-mock = "3.6.1"
pytest-mypy = "0.8.0"
pytest-xdist = {extras = ["psutil"], version = "2.3.0"}
PyYAML = "^5.4.1"
rope = "^0.20.1"
types-PyYAML = "^5.4.10"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
